Refactor code so that linear influences match general influences.
