+++ src/valuation/influence/types.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class InfluenceTypes(Enum):
+    Up = 1
+    Perturbation = 2
